+++ browser/src/libs/bitbucket/scrape.ts
@@ -87,6 +87,19 @@ export const getCommitIDFromLink = (selector = 'a.commitid'): string => {
     return commitID
 }
 
+const getCommitIDFromRevisionSelector = (): string => {
+    const revisionSelectorSpan = document.querySelector<HTMLElement>('span[data-revision-ref]')
+    if (!revisionSelectorSpan) {
+        throw new Error('Could not find span[data-revision-ref] element')
+    }
+    try {
+        const { latestCommit }: { latestCommit: string } = JSON.parse(revisionSelectorSpan.dataset.revisionRef!)
+        return latestCommit
+    } catch (err) {
+        throw new Error('Could not parse JSON from revision selector')
+    }
+}
+
 /**
  * Gets the file info on a single-file source code view
  */
@@ -94,7 +107,7 @@ export const getFileInfoFromSingleFileSourceCodeView = (
     codeViewElement: HTMLElement
 ): BitbucketRepoInfo & Pick<FileInfo, 'rawRepoName' | 'filePath' | 'rev' | 'commitID'> => {
     const { rawRepoName, filePath, rev, project, repoSlug } = getFileInfoFromLinkInSingleFileView(codeViewElement)
-    const commitID = getCommitIDFromLink()
+    const commitID = getCommitIDFromRevisionSelector()
     return {
         rawRepoName,
         filePath,
@@ -116,13 +129,14 @@ export const isCompareView = () => !!document.querySelector('#branch-compare')
 /**
  * Returns true if the active page is a commit view.
  */
-export const isCommitsView = () => /^\/projects\/[^\/]+\/repos\/[^\/]+\/commits\/\w+$/.test(window.location.pathname)
+export const isCommitsView = ({ pathname }: Pick<Location, 'pathname'>): boolean =>
+    /\/projects\/[^\/]+\/repos\/[^\/]+\/commits\/\w+$/.test(pathname)
 
 /**
  * Returns true if the active page is a pull request view.
  */
-export const isPullRequestView = () =>
-    /^\/projects\/[^\/]+\/repos\/[^\/]+\/pull-requests\/\d+/.test(window.location.pathname)
+export const isPullRequestView = ({ pathname }: Pick<Location, 'pathname'>): boolean =>
+    /\/projects\/[^\/]+\/repos\/[^\/]+\/pull-requests\/\d+/.test(pathname)
 
 /**
  * Returns true if the given code view is a single file source or "diff to previous" view.