+++ cmd/server/shared/shared.go
@@ -28,6 +28,7 @@ var defaultEnv = map[string]string{
 	"QUERY_RUNNER_URL":      "http://127.0.0.1:3183",
 	"SRC_SYNTECT_SERVER":    "http://127.0.0.1:9238",
 	"SYMBOLS_URL":           "http://127.0.0.1:3184",
+	"REPLACER_URL":          "http://127.0.0.1:3185",
 	"SRC_HTTP_ADDR":         ":8080",
 	"SRC_HTTPS_ADDR":        ":8443",
 	"SRC_FRONTEND_INTERNAL": FrontendInternalHost,
@@ -37,6 +38,7 @@ var defaultEnv = map[string]string{
 	// searcher/symbols to ensure this value isn't larger than the volume for
 	// CACHE_DIR.
 	"SEARCHER_CACHE_SIZE_MB": "50000",
+	"REPLACER_CACHE_SIZE_MB": "50000",
 	"SYMBOLS_CACHE_SIZE_MB":  "50000",
 
 	// Used to differentiate between deployments on dev, Docker, and Kubernetes.