+++ enterprise/cmd/frontend/internal/authz/bitbucketserver/provider.go
@@ -12,7 +12,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/authz"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
-	"github.com/sourcegraph/sourcegraph/pkg/api"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/bitbucketserver"
 	"github.com/sourcegraph/sourcegraph/pkg/trace"
@@ -65,15 +64,13 @@ func (p *Provider) ServiceID() string { return p.codeHost.ServiceID }
 // ServiceType returns the type of this Provider, namely, "bitbucketServer".
 func (p *Provider) ServiceType() string { return p.codeHost.ServiceType }
 
-// Repos returns all Bitbucket Server repos as mine, and all others as others.
-func (p *Provider) Repos(ctx context.Context, repos map[authz.Repo]struct{}) (mine map[authz.Repo]struct{}, others map[authz.Repo]struct{}) {
-	return authz.GetCodeHostRepos(p.codeHost, repos)
-}
-
 // RepoPerms returns the permissions the given external account has in relation to the given set of repos.
 // It performs a single HTTP request against the Bitbucket Server API which returns all repositories
 // the authenticated user has permissions to read.
-func (p *Provider) RepoPerms(ctx context.Context, acct *extsvc.ExternalAccount, repos map[authz.Repo]struct{}) (authorized map[api.RepoName]map[authz.Perm]bool, err error) {
+func (p *Provider) RepoPerms(ctx context.Context, acct *extsvc.ExternalAccount, repos []*types.Repo) (
+	perms []authz.RepoPerms,
+	err error,
+) {
 	var (
 		userName string
 		userID   int32
@@ -85,7 +82,7 @@ func (p *Provider) RepoPerms(ctx context.Context, acct *extsvc.ExternalAccount, 
 			otlog.String("user.name", userName),
 			otlog.Int32("user.id", userID),
 			otlog.Int("repos.count", len(repos)),
-			otlog.Int("authorized.count", len(authorized)),
+			otlog.Int("perms.count", len(perms)),
 		)
 
 		if err != nil {
@@ -105,9 +102,9 @@ func (p *Provider) RepoPerms(ctx context.Context, acct *extsvc.ExternalAccount, 
 		userName = user.Name
 	}
 
-	ids := make(map[int]authz.Repo, len(repos))
-	for r := range repos {
-		if id, _ := strconv.Atoi(r.ExternalRepoSpec.ID); id != 0 {
+	ids := make(map[int]*types.Repo, len(repos))
+	for _, r := range repos {
+		if id, _ := strconv.Atoi(r.ExternalRepo.ID); id != 0 {
 			ids[id] = r
 		}
 	}